with services after remarks

+++ src/main/java/com/example/flashcardsv4/service/CardServiceImpl.java
@@ -4,35 +4,46 @@ package com.example.flashcardsv4.service;
 import com.example.flashcardsv4.exceptions.InputDataException;
 import com.example.flashcardsv4.models.Card;
 import com.example.flashcardsv4.repositories.CardRepository;
+import com.example.flashcardsv4.repositories.ChapterRepository;
 
 import java.util.List;
 
 public class CardServiceImpl implements CardService {
     CardRepository cardRepository;
+    ChapterRepository chapterRepository;
 
     public CardServiceImpl(CardRepository cardRepository) {
         this.cardRepository = cardRepository;
     }
 
     public void deleteCard(long cardId) {
-        cardRepository.deleteCard(cardId);
-    }
-
-    public void addCard(long chapterId, String question, String answer, boolean isRemembered) {
-        if (!question.isEmpty() && !answer.isEmpty()) {
-            cardRepository.addCard(chapterId, question, answer, isRemembered);
+        if (cardRepository.ifCardExists(cardId)) {
+            cardRepository.deleteCard(cardId);
         } else throw new InputDataException();
     }
 
-    public void updateMemorizing(long cardId, long chapterId) {
-        cardRepository.updateMemorizing(cardId);
-        cardRepository.getOneCardData(chapterId, cardId);
+    public void addCard(long chapterId, String question, String answer) {
+        if (chapterRepository.ifChapterExists(chapterId)) {
+            if (!question.isEmpty() && !answer.isEmpty()) {
+                cardRepository.addCard(chapterId, question, answer);
+            } else throw new InputDataException();
+        } else throw new InputDataException();
     }
 
-    public List<Card> getOneCardData(long chapterId, long offset) {
-        return cardRepository.getOneCardData(chapterId, offset);
+    @Override
+    public void updateMemorizingToTrueForTraining(long cardId) {
+        if (cardRepository.ifCardExists(cardId)) {
+            cardRepository.updateMemorizing(cardId);
+        } else throw new InputDataException();
     }
-    public List<Card> showAllCards(long chapterId){
+
+    public List<Card> getOneCardDataForTraining(long chapterId, long offset) {
+        if (chapterRepository.ifChapterExists(chapterId)) {
+            return cardRepository.getOneCardData(chapterId, offset);
+        } else throw new InputDataException();
+    }
+
+    public List<Card> showAllCards(long chapterId) {
         return cardRepository.showAllCards(chapterId);
     }
 }
